# Backtesting speed optimizations