This PR makes two improvements to backtesting speed:
- Faster date range printing
- Less data passing in the core loop

First is 136x faster than the original, on my laptop it reduces roughly 7 seconds off from backtest run.

Second picks just the columns we need from the dataframe before converting them to tuples. This makes the tuples faster to create and reduces garbage collection too.

The whole backtesting run on my laptop went from:

`python freqtrade/main.py backtesting  76.48s user 0.48s system 98% cpu 1:17.90 total`

to

`python freqtrade/main.py backtesting  51.60s user 0.91s system 98% cpu 53.067 total`

Total time was reduced by 32%.