@glonlas 2018-01-03
LGTM 👍  Impatient to experiment it.